% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{caculate_llh}
\alias{caculate_llh}
\title{This function finds the best K of an estimated GARCH-MIDAS model}
\usage{
caculate_llh(data, y, x, freq = "month", K.seq = c(6:48))
}
\arguments{
\item{data}{data frame containing a column named date of type 'Date'.}

\item{y}{name of high frequency dependent variable in df.}

\item{x}{covariate employed in GarchMidas.}

\item{freq}{a string of the low frequency variable in the df.}

\item{K.seq}{an integer sequence specifying lag length K in the long-term component.}
}
\description{
This function finds the best K of an estimated GARCH-MIDAS model
}
