% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fu}
\alias{fu}
\title{Stock returns and uncertainty index.}
\format{A data frame with 3,789 rows and 7 variables:
\describe{
  \item{date}{date}
  \item{return}{daily SSE log returns}
  \item{hundredfold_return}{daily SSE log returns times 100}
  \item{rv}{5-minute realized variances}
  \item{month}{a dummy for each year/month combination}
  \item{fu}{Financial Uncertainty Index}
  \item{hundredfold_fu}{Financial Uncertainty Index times 100}
}}
\usage{
fu
}
\description{
A dataset containing SSE (Shanghai Stock Exchange) Composite Index returns and the China Financial Uncertainty Index
}
\keyword{datasets}
