% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCS.R
\name{MCS}
\alias{MCS}
\title{This function finds the Model Confidence Set}
\usage{
MCS(Loss, alpha = 0.1, B = 5000, statistic = "Tmax",
  verbose = TRUE)
}
\arguments{
\item{Loss}{data frame of loss function}

\item{alpha}{confidence level}

\item{B}{bootstrap times}

\item{statistic}{"Tmax" or "TR"}
}
\description{
This function finds the Model Confidence Set
}
