% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mu}
\alias{mu}
\title{Stock returns and uncertainty index.}
\format{A data frame with 3,789 rows and 5 variables:
\describe{
  \item{date}{date}
  \item{return}{daily SSE log returns times 100}
  \item{rv}{5-minute realized variances}
  \item{month}{a dummy for each year/month combination}
  \item{mu}{Macroeconomic Uncertainty Index}
}}
\usage{
mu
}
\description{
A dataset containing SSE (Shanghai Stock Exchange) Composite Index returns and the Macroeconomic Uncertainty Index
}
\keyword{datasets}
