% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_GarchMidas.R
\name{fit_GarchMidas}
\alias{fit_GarchMidas}
\title{This function estimates a multiplicative mixed-frequency GARCH model. For the sake of numerical stability, it is best to multiply log returns by 100.}
\usage{
fit_GarchMidas(data, y, x, K, freq = "month", GJR = FALSE)
}
\arguments{
\item{data}{data frame containing a column named date of type 'Date'.}

\item{y}{name of high frequency dependent variable in df.}

\item{x}{covariate employed in GarchMidas.}

\item{K}{an integer specifying lag length K in the long-term component.}

\item{freq}{a string of the low frequency variable in the df.}

\item{GJR}{if TRUE, an asymmetric GJR-GARCH is used as the short-term component. If FALSE, a simple GARCH(1,1) is employed.}
}
\description{
This function estimates a multiplicative mixed-frequency GARCH model. For the sake of numerical stability, it is best to multiply log returns by 100.
}
\examples{
\dontrun{
fit_GarchMidas(data = mu, y = "return", x = "epu", K = 24, freq = "month", GJR = TRUE)
}
}
\keyword{fit_GarchMidas}
