% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_GarchMidas.R
\name{fit_GarchMidas}
\alias{fit_GarchMidas}
\title{This function estimates a multiplicative mixed-frequency GARCH model. For the sake of numerical stability, it is best to multiply log returns by 100.}
\usage{
fit_GarchMidas(data, y, x, K, low.freq = "month",
  var.ratio.freq = NULL)
}
\arguments{
\item{data}{data frame containing a column named date of type 'Date'.}

\item{y}{name of high frequency dependent variable in df.}

\item{x}{covariate employed in GarchMidas.}

\item{K}{an integer specifying lag length K in the long-term component.}

\item{low.freq}{a string of the low frequency variable in the df.}

\item{var.ratio.freq}{specify a frequency column on which the variance ratio should be calculated.}
}
\description{
This function estimates a multiplicative mixed-frequency GARCH model. For the sake of numerical stability, it is best to multiply log returns by 100.
}
\examples{
\dontrun{
fit_GarchMidas(data = mu, y = "return", x = "mu", low.freq = "month", K = 24)
}
}
\keyword{fit_GarchMidas}
